fix #24 : ProblemId Integer -> Long 변경

+++ backend/problem/src/main/java/com/scf/problem/domain/dto/ProblemContentDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class ProblemContentDTO {
 
-    private Integer problemId;
+    private Long problemId;
 
     private String content;
